changes to acq obj saving in motionCorrect
motionCorrect now checks to see if the defaultDir is empty (which it
only is permitted to be if a writeDir is specified). If defaultDir is
empty, the writeDir is assigned as the defaultDir. This is important
because at the end of motion correction the acquisition object is now
saved in the object’s defaultDir. This is nice aesthetically, and also
allows easier/automatic overwriting by downstream functions by just
saving the object to its own defaultDir