Updated extractRois for new binary movie files.
Use extractROIsBin with binary movie files, or extractROIsTIFF for tiff
files. The old extractROIs was renamed into extractROIsOld for use with
existing binary files. This means that code still using the old bin file
will be automatically broken (no mor extractROIs function) so that the
spots that need to be updated can be found quickly.