Job processor now saves in between steps
Updated the job processor to save acquisition object to inProgress
directory and default directory in between steps. Additionally, the job
processor now checks if a step has been already completed before
performing it.  So, for example, if you stopped the job processor in the
middle of pixel covariance calculation, and moved the file back from
inProgress into the queue, it would skip motion correction and binary
file creation for that file and only perform pixel covariance
calculation. It would, however have to redo everything for pixel
covariance.