Rewrote selectROIs as class. Update to pxCov and binMov.
The selectROIs GUI was completely re-written. Changes to functionality
are minimal, but the code is very different. Two changes affect
compatibility: selectROIs now requires a binary movie file (evaluating
traces is necessary for judging ROIs). selectROIs also now requires the
pixel covariances to be calculated with the function written by Matthias
(calcPxCovMJLM, but to be renamed to calcPcCov).