Changed ROI data structure.
Changed ROI data structure to be safer to maintain. Information specific
to a particular ROI is now all stored in an array of structures with
nRoi elements and the fields "id", "group", "indBody", "indNeuropil" and
"subCoef". ID is a unique identifier for a given ROI that is never
re-used even if the ROI is deleted. selectROIs and extractROIsBin have
been updated, and a conversion function for existing acquisitions has
been added to the auxiliary function folder.