Removed neuropil index sub-selection
Empirically did not appear useful anymore. Often caused slight bias in
fitting of neuropil coef, so that very high neuropil values caused
slight negative deflections in trace.

This change is not so important, as eventually the fit will be user
adjustable, but I wanted to eliminate this behavior in the 'default'
case