New acq property for metadata, size moved to corMov
Acq objects have a new metadata property that is filled during motion
correction. Movie sizes were also moved from the derivedData field
(which really should be for non-essential info), to an array in
correctedMovies. For simplicity / compatability with the correctedMovies
field format, and because it is conceivable at some future timepoint,
each slice and each channel receives its own movie size array. There is
no disadvantage if this proves redundant since the storage space is
miniscule.