Scroll-Wheel neuropil scatter plot adjust of fit
Broke out some functions from plotNeuropilTraces to avoid redundancy.
New scroll wheel callback adjusts the fit's slope, and replots the line
(in scatter figure) and dF-sub trace (in neuropil-sub figure) in
realtime. It seemed superfluous to delay or use a timer for either of
these, at least on my computer.