'Raw Plot' checkbox to plot dF w/out neuropil-sub
Small checkbox placed (automatically positioned, I haven't specified
this at all) on neuropil-sub traces figure, that dictates whether to
plut the raw in addition to the subtracted dF/F. The checkbox doesn't
trigger a callback, but the neuropil fit scroll wheel causes a re-plot,
so that's an easy way to toggle the raw plot option if desired.

For clarity, the sub trace is always plotted second, on top of the raw
trace. I thus modified the plot so that the sub trace is always the
orange-red color it gets when plotted second (after the raw), even when
it's the only trace plotted