added Overlay Trace w/ 't' keypress
currently displays raw mean pixel values. This seems most useful for
uses I've imagined, but could be up for debate. Idea is that it's useful
to see the unbleached, un-normalized values somewhere, it's useful to
compare relative intensities to see how much brighter an ROI is than
background, and when comparing two different 'cuts' of the same ROI it
doesn't really matter if the trace is normalized or not