Add automatic backup to cell selection
A file with the ROI-info is saved to acq.defaultDir every time an ROI is
selected. The amount of saved data is small so this is unnoticeably
fast. If the cell selection too crashes or the user forgets to save, the
work can be restored by re-loading the acq file and overwriting its
(outdated) ROI info with that from the backup file.