In getCovData, allow user to query a sub-part of the full neighborhood
I (Matthias) made a function that calculates a new kind of "activity
overview image" which highlights very small, highly active sources (e.g.
dendrites). This makes it much easier to select such sources. To
calculate this overview image, I use the covData in a 3-by-3
neighborhood. To speed up the calculation, I added this functionality to
the getCovData function.