Fix bug in syncInfo creation that caused only one movie to be added to syncInfo
SyncInfo assumed that the movie list is a row vector, but that is not
enforced and it could as well be a column vector. In that case, all
filenames were concatenated into one long, nonsensical string. This is
fixed here by using cellfun instead of concatenating the vector into a
character matrix.